requirements.txt
appdirs==1.4.3
Django==1.8
django-allauth==0.22.0
django-filter==1.0.2
djangorestframework==3.6.2
Markdown==2.6.8
MySQL-python==1.2.5
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0